requirements.txt
playwright==1.50.0
samsungtvws[async]==3.0.4
jinja2==3.1.5
httpx==0.28.1
pyyaml==6.0.2
pillow==11.1.0
